[X86] Teach combineToExtendBoolVectorInReg to create opportunities for using broadcast load instructions.
If we're inserting a scalar that is smaller than the element
size of the final VT, the value of the extra bits doesn't matter.

Previously we any_extended in the scalar domain before inserting.

This patch changes this to use a broadcast of the original
scalar type and then a bitcast to the final type. This might
enable the use of a broadcast load.

This recovers regressions from 632e8b7999c73391959276e93ac00dc50d13ffe7
and 60f3c3b74149385a63c0d53ced0db907696fa656 without relying on
alignment of the load.

Differential Revision: https://reviews.llvm.org/D75835