[PowerPC][Future] Fix for 2b1e6414b0d3e25f7ce61d300b8d6fa858f88cf4
This is a fix for the previous patch 2b1e6414b0d3e25f7ce61d300b8d6fa858f88cf4.
In some cases it may be possible to have the compiler produce st_other=1 without
the compiler using mcpu=future which should not be the case. This patch adds a
guard to make sure that if we are using st_other=1 then we are also compiling
for future CPU.